breakout mod GreyScale adaptor iterator
